requirements.txt
# streamlit
# opencv-python-headless
# numpy
# onnxruntime
# scikit-learn
# Pillow




streamlit
opencv-python-headless==4.10.0.84
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.6.1
onnxruntime
Pillow
joblib
